Work around the gcc warning in array.c again
Despite the warning related to an always-false condition, gcc
was not optimising away the check for from > to (but maybe it
optimised away one of the other checks?)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8225 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ array.c
@@ -577,7 +577,8 @@ array_t array_delete_slice(array_t *array, int from, int to) {
 	typetable *tytbl = get_type(*array);
 	unsigned int len = length(*array);
 
-	if (from < 0 || (to > len || from > to)) {
+	// Cast is just to silence an annoying warning due to a gcc bug
+	if (from < 0 || to > len || (unsigned int)from > (unsigned int)to) {
 		debug(errPromptBug, "array_delete_slice: invalid slice [%d, %d) of array of length %d", from, to, len);
 		return *array;
 	}
